# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Logistics Enquiry Triage Tool",
    "version": "0.1.0"
  },
  "paths": {
...
    "/enquiries": {
      "post": {
        "tags": [
          "enquiries"
        ],
        "summary": "Ingest Enquiry",
        "description": "Ingest a new enquiry (A4). Computes code warnings on insert.",
        "operationId": "ingest_enquiry_enquiries_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnquiryIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnquiryOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "enquiries"
        ],
        "summary": "List Enquiries",
        "description": "List enquiries with filters, full-text search, and cursor pagination (\u00a74.2).",
        "operationId": "list_enquiries_enquiries_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "company",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Company"
            }
          },
          {
            "name": "priority",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by latest analysis priority",
              "title": "Priority"
            },
            "description": "Filter by latest analysis priority"
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by latest analysis category",
              "title": "Category"
            },
            "description": "Filter by latest analysis category"
          },
          {
            "name": "safety_verdict",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by latest safety verdict",
              "title": "Safety Verdict"
            },
            "description": "Filter by latest safety verdict"
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Full-text search",
              "title": "Q"
            },
            "description": "Full-text search"
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "received_at | company",
              "default": "received_at",
              "title": "Sort"
            },
            "description": "received_at | company"
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Keyset cursor (received_at ISO of last item)",
              "title": "Cursor"
            },
            "description": "Keyset cursor (received_at ISO of last item)"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnquiryListOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/enquiries/{id}": {
      "get": {
        "tags": [
          "enquiries"
        ],
        "summary": "Get Enquiry",
        "operationId": "get_enquiry_enquiries__id__get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnquiryOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EnquiryIn": {
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 200,
            "title": "Id",
            "description": "External id from ingest"
          },
          "name": {
            "type": "string",
            "maxLength": 200,
            "title": "Name",
            "default": ""
          },
          "email": {
            "type": "string",
            "maxLength": 200,
            "title": "Email"
          },
          "company": {
            "type": "string",
            "maxLength": 200,
            "title": "Company"
          },
          "status": {
            "type": "string",
            "maxLength": 100,
            "title": "Status",
            "description": "Source status"
          },
          "message": {
            "type": "string",
            "maxLength": 5000,
            "title": "Message"
          },
          "received_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Received At",
            "description": "Defaults to now() if omitted"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "company",
          "status",
          "message"
        ],
        "title": "EnquiryIn",
        "description": "Ingest payload \u2014 matches the `enquiries` table's external fields."
      },
      "EnquiryListOut": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/EnquiryOut"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "type": "object",
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "EnquiryListOut",
        "description": "Response shape for GET /enquiries (\u00a74.2: {items, next_cursor})."
      },
      "EnquiryOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "company": {
            "type": "string",
            "title": "Company"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "received_at": {
            "type": "string",
            "format": "date-time",
            "title": "Received At"
          },
          "warning_missing_name": {
            "type": "boolean",
            "title": "Warning Missing Name"
          },
          "warning_invalid_email": {
            "type": "boolean",
            "title": "Warning Invalid Email"
          },
          "warning_domain_mismatch": {
            "type": "boolean",
            "title": "Warning Domain Mismatch"
          },
          "warning_possible_duplicate": {
            "type": "boolean",
            "title": "Warning Possible Duplicate"
          },
          "duplicate_of": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duplicate Of"
          },
          "claimed_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Claimed By"
          },
          "claim_expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Claim Expires At"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "latest_analysis_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Analysis Id"
          },
          "analysis_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analysis Status"
          },
          "analysis_result": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analysis Result"
          },
          "safety_decision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Safety Decision"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "email",
          "company",
          "status",
          "message",
          "received_at",
          "warning_missing_name",
          "warning_invalid_email",
          "warning_domain_mismatch",
          "warning_possible_duplicate",
          "duplicate_of",
          "claimed_by",
          "claim_expires_at",
          "version",
          "created_at",
          "updated_at"
        ],
        "title": "EnquiryOut",
        "description": "Response shape for GET /enquiries, GET /enquiries/{id}."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}